move FAPI callback constants to internal/constants
The callback types are used in scripts/prepare_headers.py so they
can not be in a module that imports ._libtpm2_pytss
