+++ scripts/prepare_headers.py
@@ -13,7 +13,7 @@ import textwrap
 
 # import tpm2_pytss.constants
 constants_spec = importlib.util.spec_from_file_location(
-    "tpm2_pytss.constants", "tpm2_pytss/constants.py"
+    "tpm2_pytss.internal.constants", "tpm2_pytss/internal/constants.py"
 )
 constants = importlib.util.module_from_spec(constants_spec)
 constants_spec.loader.exec_module(constants)
