+++ setup.cfg
@@ -20,7 +20,9 @@ classifiers =
     Programming Language :: Python :: 3.9
 
 [options]
-packages = find:
+packages =
+    tpm2_pytss
+    tpm2_pytss/internal
 setup_requires =
     setuptools_scm[toml]>=3.4.3
     cffi>=1.0.0
