+++ tpm2_pytss/callbacks.py
@@ -4,7 +4,7 @@ SPDX-License-Identifier: BSD-2
 
 
 from ._libtpm2_pytss import lib
-from .constants import CALLBACK_BASE_NAME, CALLBACK_COUNT, CallbackType
+from .internal.constants import CALLBACK_BASE_NAME, CALLBACK_COUNT, CallbackType
 
 
 class Callback:
